language/requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
aniso8601==9.0.1
ansible==4.3.0
ansible-core==2.11.3
antlr4-python3-runtime==4.9.3
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrdict==2.0.1
attrs==19.3.0
audioread==2.1.9
Automat==0.8.0
Babel==2.10.1
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
benepar==0.2.0
black==19.10b0
bleach==5.0.0
blinker==1.4
blis==0.7.7
boto3==1.23.6
botocore==1.26.7
bottle==0.12.15
braceexpand==0.1.7
breezy==3.0.2
Brlapi==0.7.0
cachetools==5.0.0
cajarename==19.7.15
catalogue==2.0.7
catkin-pkg-modules==0.4.24
cca-zoo==1.10.16
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
charset-normalizer==2.0.12
chrome-gnome-shell==0.0.0
click==8.1.3
cloud-init==21.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contextualSpellCheck==0.4.1
crit==0.0.1
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.6
Cython==0.29.30
dbus-python==1.2.16
debugpy==1.6.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
deja-dup-caja==0.0.6
Deprecated==1.2.7
Distance==0.1.3
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
docker==4.1.0
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.16
dulwich==0.19.15
duplicity==0.8.12.0
editdistance==0.5.3
empy==3.3.2
en-core-web-md==3.3.0
en-core-web-sm==3.3.0
entrypoints==0.3
executing==0.8.3
fasteners==0.14.1
fastimport==0.9.8
fastjsonschema==2.15.3
fasttext==0.9.2
filelock==3.0.12
Flask==2.1.2
Flask-RESTful==0.3.9
flatbuffers==2.0
folder-color-caja==0.0.86
folder-color-common==0.0.86
fonttools==4.33.3
frozendict==2.3.2
frozenlist==1.3.0
fsspec==2022.2.0
ftfy==6.1.1
future==0.18.2
g2p-en==2.1.0
gast==0.5.3
gdown==4.4.0
gencpp==0.6.2
genlisp==0.4.17
genmsg==0.5.14
genpy==0.6.9
gitdb==4.0.9
GitPython==3.1.27
Glances==3.1.3
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpg===1.13.1-unknown
grpcio==1.44.0
gssapi==1.6.1
h5py==3.6.0
html5lib==1.0.1
httplib2==0.14.0
huggingface-hub==0.7.0
hydra-core==1.2.0
hyperlink==19.0.0
idna==2.8
imagesize==1.3.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
incremental==16.10.1
inflect==5.6.0
influxdb==5.2.0
iniconfig==1.1.1
iotop==0.6
ipaclient==4.8.6
ipadic==1.0.0
ipalib==4.8.6
ipaplatform==4.8.6
ipapython==4.8.6
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isc==2.0
isort==4.3.21
itsdangerous==2.1.2
jarowinkler==1.0.2
jedi==0.18.1
jieba==0.42.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==1.1.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kaldi-io==0.9.4
kaldi-python-io==1.2.2
kaldiio==2.17.2
kazam==1.4.5
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.0.1
langcodes==3.3.0
language-selector==0.1
latexcodec==2.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==13.0.0
librosa==0.9.1
llvmlite==0.38.1
lmdb==1.3.0
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Magnus==1.0.3
Mako==1.1.0
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.15.0
mate-hud==19.10.0
mate-menu==20.4.1
mate-tweak==20.4.0
matplotlib==3.5.2
matplotlib-inline==0.1.3
mecab-python3==1.0.5
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
mpmath==1.2.1
multidict==6.0.2
murmurhash==1.0.7
mvlearn==0.4.1
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nemo-toolkit==1.8.2
nest-asyncio==1.5.5
netaddr==0.7.19
netifaces==0.10.4
nltk==3.7
notebook==6.4.11
numba==0.55.1
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
omegaconf==2.2.1
onboard==1.4.1
onnx==1.11.0
OpenCC==1.1.3
opt-einsum==3.3.0
packaging==21.0
pandas==1.4.1
pandocfilters==1.5.0
pangu==4.0.6.1
parameterized==0.8.1
paramiko==2.6.0
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pathy==0.6.1
pdfarranger==1.4.2
pesq==0.0.4
pexpect==4.6.0
pickleshare==0.7.5
pikepdf==1.10.3+dfsg
Pillow==7.0.0
pip-api==0.0.29
pipreqs==0.4.11
pluggy==1.0.0
ply==3.11
pooch==1.6.0
portalocker==2.4.0
preshed==3.0.6
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.5.1
ptyprocess==0.7.0
pulsemixer==1.5.0
pure-eval==0.2.2
py==1.11.0
pyannote.core==4.4
pyannote.database==4.1.3
pyannote.metrics==3.2
pyarrow==7.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pybind11==2.9.2
pybtex==0.24.0
pybtex-docutils==1.0.1
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycryptodomex==3.6.1
pycups==1.9.73
pydantic==1.8.2
pyDeprecate==0.3.2
pydub==0.25.1
PyGithub==1.43.7
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynini==2.1.4
pyOpenSSL==19.0.0
pyparsing==2.4.7
pypinyin==0.46.0
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
pysmi==0.3.2
pysnmp==4.4.6
pystache==0.5.4
pystoi==0.3.3
pytest==7.1.2
pytest-runner==6.0.0
python-apt==2.0.0+ubuntu0.20.4.7
python-augeas==0.5.0
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-gitlab==2.0.1
python-gnupg==0.4.5
python-ldap==3.2.0
python-qt-binding==0.3.6
python-xapp==1.8.1
python-xlib==0.23
python-yubico==1.3.3
pytorch-lightning==1.6.0
pytz==2022.1
pyudev==0.21.0
pyusb==1.0.2
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.4.1
pyzmq==23.0.0
qrcode==6.1
rapidfuzz==2.0.11
regex==2022.3.15
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
resampy==0.2.2
resolvelib==0.5.4
roman==2.0.0
rospkg==1.4.0
rospkg-modules==1.4.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
rubber==1.5.1
s3transfer==0.5.2
sacrebleu==2.1.0
sacremoses==0.0.53
scikit-learn==1.0.2
scipy==1.8.0
screen-resolution-extra==0.0.0
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.8.0
sentencepiece==0.1.96
sentry-sdk==1.5.12
service-identity==18.1.0
setproctitle==1.1.10
shellingham==1.4.0
shortuuid==1.0.9
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sos==4.3
SoundFile==0.10.3.post1
soupsieve==1.9.5
sox==1.4.1
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-bibtex==2.4.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
srsly==2.4.3
ssh-import-id==5.10
stack-data==0.2.0
sympy==1.10.1
systemd-python==234
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
tensorly==0.7.0
termcolor==1.1.0
terminado==0.15.0
textblob==0.17.1
tf-estimator-nightly==2.8.0.dev2021122109
thinc==8.0.16
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
torch==1.11.0+cu113
torch-stft==0.1.4
torch-struct==0.5
torchaudio==0.11.0+cu113
torchmetrics==0.7.3
torchvision==0.12.0+cu113
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
transformers==4.19.2
Twisted==18.9.0
typed-ast==1.5.4
typer==0.4.1
typing-extensions==4.1.1
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.3.4
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
wandb==0.12.16
wasabi==0.9.1
wcwidth==0.2.5
webdataset==0.1.62
webencodings==0.5.1
websocket-client==0.53.0
Werkzeug==2.1.0
wget==3.2
widgetsnbextension==3.6.0
wrapt==1.11.2
xkit==0.0.0
yarg==0.1.9
yarl==1.7.2
youtokentome==1.0.6
youtube-dl==2020.3.24
zipp==3.8.0
zope.interface==4.7.1
